update: 警告の抑制

+++ src/description_field_renderer.h
@@ -32,7 +32,7 @@ class DescriptionFieldRenderer final {
 
   void Update();
 
-  void Draw(int x, int y) const;
+  void Draw(float x, float y) const;
 
  private:
   std::string GetString() const;
