+++ src/score_board_renderer.h
@@ -34,7 +34,7 @@ class ScoreBoardRenderer final {
     tetris_game_mode_ = game_mode;
   }
 
-  void Draw(int x, int y) const;
+  void Draw(float x, float y) const;
 
  private:
   std::string GetString() const;
